+++ gd-yd-core/src/main/java/com/gdyd/gdydcore/service/mentoring/HighSchoolStudentQuestionService.java
@@ -28,7 +28,7 @@ public class HighSchoolStudentQuestionService {
         return highSchoolStudentQuestionRepository.findAll(pageable);
     }
 
-    public List<HighSchoolStudentQuestion> findByLikeCountGreaterThanEqualAndCreatedAtAfter(int like, LocalDateTime weeksAgo, Pageable pageable) {
+    public List<HighSchoolStudentQuestion> findByLikeCountGreaterThanEqualAndCreatedAtAfter(Long like, LocalDateTime weeksAgo, Pageable pageable) {
         return highSchoolStudentQuestionRepository.findByLikeCountGreaterThanEqualAndCreatedAtAfter(like, weeksAgo, pageable);
     }
